refactor: move server stats provider and register to internal
